beautify.js, process.js, sort.js統合
